♻️ Refactor: WIP, grouping 해제 및 layout&page 작업
