+++ src/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)search/page.tsx
