+++ src/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)settings/(.)change-password/default.tsx
