+++ src/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)settings/(.)change-password/page.tsx
